pyproject.toml
[build-system]
requires = [ "poetry>=1.1.0" ]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "pipen"
version = "0.15.8"
description = "A pipeline framework for python"
authors = [ "pwwang <pwwang@pwwang.com>",]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/pwwang/pipen"
repository = "https://github.com/pwwang/pipen"

[tool.poetry.build]
generate-setup-file = true

# fix ssl certificate issue on gcp head nodes
[[tool.poetry.source]]
name = "fpho"
url = "https://files.pythonhosted.org"
priority = "primary"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.dependencies]
python = "^3.9"
liquidpy = "^0.8.3"
pandas = "^2.0"
enlighten = "^1"
argx = "^0.3"
# xqute = "^0.8"
xqute = { git = "https://github.com/pwwang/xqute.git" }
## included in xqute
# rich = "^13"
# diot = "^0.1"
# simplug = "^0.0"
python-simpleconf = {version = "^0.6.2", extras = ["toml"]}
pipda = "^0.13"
varname = "^0.14"

[tool.poetry.group.dev.dependencies]
openpyxl = "^3"
pytest = "^8"
pytest-asyncio = "^0.25"
pytest-cov = "^6"
pytest-xdist = "^3"
pytest-forked = "^1.6"
# This also installs scipy and wcwidth
datar = { version = "^0.15", extras = ["pandas"] }
flake8 = "^7.1.2"
python-dotenv = "^1"
cloudsh = {version = "^0.1", extras = ["gs"]}

# dependencies for pipelines in ./examples
[tool.poetry.group.example.dependencies]
mako = "^1.3"
python-dotenv = "^1"
cloudsh = {version = "^0.1", extras = ["gs"]}

[tool.poetry.scripts]
pipen = "pipen.cli:main"

[tool.pytest.ini_options]
addopts = "-vv -n auto --dist loadgroup -p no:benchmark -W error::UserWarning --cov-config=.coveragerc --cov=pipen --cov-report xml:.coverage.xml --cov-report term-missing"
console_output_style = "progress"
junit_family = "xunit1"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
   "ignore:.+may lead to deadlocks in the child:DeprecationWarning",
]

[tool.mypy]
ignore_missing_imports = true
allow_redefinition = true
disable_error_code = ["attr-defined", "no-redef"]
show_error_codes = true
strict_optional = false

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
